Rename discord-notifier
