+++ src/bot/events/EntitlementDelete.ts
@@ -1,7 +1,7 @@
 import { Entitlement } from "discord.js";
 import Event from "../structs/Event";
 import players from "../../database/schemas/players";
-import { sendEntitlementMessage } from "../../libs/DiscordNotifier";
+import { sendEntitlementMessage } from "../../libs/discord-notifier";
 import entitlements from "../../database/schemas/entitlement";
 import { config } from "../../libs/Config";
 import { getSkus } from "../../libs/SkuManager";
